requirements.txt
astroid==2.2.5
backcall==0.1.0
colorama==0.4.1
decorator==4.3.2
Django==2.1.7
django-cors-headers==2.4.0
django-filter==2.1.0
djangorestframework==3.9.1
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
isort==4.3.13
jedi==0.13.3
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
parso==0.3.4
pep8==1.7.1
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.7.7
Pygments==2.3.1
pylint==2.3.1
pytz==2018.9
six==1.12.0
traitlets==4.3.2
typed-ast==1.3.1
wcwidth==0.1.7
wrapt==1.11.1
dj-database-url==0.5.0
gunicorn==19.9.0
psycopg2-binary==2.7.7
whitenoise==4.1.2